requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py38h497a2fe_0
ase=3.21.1=pypi_0
astroid=2.5.6=py38h578d9bd_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
azure-common=1.1.27=pyhd8ed1ab_0
azure-core=1.12.0=pyhd8ed1ab_0
azure-storage-blob=12.8.1=pyhd8ed1ab_0
blas=1.0=mkl
blessings=1.7=py38h578d9bd_1004
blinker=1.4=py_1
boto=2.49.0=py_0
boto3=1.17.72=pyhd8ed1ab_0
botocore=1.20.72=pyhd8ed1ab_0
brotlipy=0.7.0=py38h497a2fe_1001
c-ares=1.17.1=h7f98852_1
ca-certificates=2020.12.5=ha878542_0
cachetools=4.2.2=pyhd8ed1ab_0
certifi=2020.12.5=py38h578d9bd_1
cffi=1.14.5=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
click=8.0.0=py38h578d9bd_0
cryptography=3.4.7=py38ha5dfef3_0
cudatoolkit=10.2.89=h8f6ccaa_8
cycler=0.10.0=py_2
dataclasses=0.8=pyhc8e2a94_1
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
expat=2.3.0=h9c3ff4c_0
flake8=3.9.2=pyhd8ed1ab_0
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gensim=4.0.1=py38h709712a_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.2=h9c3ff4c_0
glib-tools=2.68.2=h9c3ff4c_0
google-api-core=1.26.3=pyhd8ed1ab_0
google-auth=1.30.0=pyh44b312d_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-core=1.5.0=pyhd3deb0d_0
google-cloud-storage=1.19.0=py_0
google-crc32c=1.1.2=py38h8838a9a_0
google-resumable-media=1.2.0=pyhd3deb0d_0
googleapis-common-protos=1.53.0=py38h578d9bd_0
googledrivedownloader=0.4=pypi_0
gpustat=0.6.0=pyhd8ed1ab_1
grpcio=1.37.1=py38hdd6454d_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
h5py=3.2.1=pypi_0
icu=58.2=hf484d3e_1000
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.0.1=py38h578d9bd_0
intel-openmp=2021.2.0=h06a4308_610
isodate=0.6.0=py_1
isort=5.8.0=pyhd8ed1ab_0
jinja2=3.0.0=pypi_0
jmespath=0.10.0=pyh9f0ad1d_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonlines=2.0.0=pyhd3deb0d_0
kiwisolver=1.3.1=py38h1fd1430_1
lazy-object-proxy=1.6.0=py38h497a2fe_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libcrc32c=1.1.1=h9c3ff4c_2
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=9.3.0=hff62375_19
libgfortran5=9.3.0=hff62375_19
libglib=2.68.2=h3e27bee_0
libgomp=9.3.0=h2828fa1_19
libiconv=1.16=h516909a_0
liblapack=3.9.0=9_mkl
libopenblas=0.3.15=pthreads_h8fe5266_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.16.0=h780b84a_0
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.2.0=hdc55705_1
libuuid=2.32.1=h7f98852_1000
libuv=1.41.0=h7f98852_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.0=pypi_0
matplotlib=3.4.2=py38h578d9bd_0
matplotlib-base=3.4.2=py38hcc49a3a_0
mccabe=0.6.1=py_1
mkl=2021.2.0=h06a4308_296
msrest=0.6.21=pyh44b312d_0
multidict=5.1.0=py38h497a2fe_1
ncurses=6.2=h58526e2_4
networkx=2.5.1=pyhd8ed1ab_0
ninja=1.10.2=h4bd325d_0
nltk=3.6.2=pyhd8ed1ab_0
numba=0.53.1=pypi_0
numpy=1.20.2=py38h9894fe3_0
nvidia-ml=7.352.0=py_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hf7af979_0
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=pypi_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pillow=8.2.0=pypi_0
pip=21.1.1=pyhd8ed1ab_0
protobuf=3.16.0=py38h709712a_0
psutil=5.8.0=py38h497a2fe_1
pthread-stubs=0.4=h36c2ea0_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.7.0=pyhd8ed1ab_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.3.1=pyhd8ed1ab_0
pyjwt=2.1.0=pyhd8ed1ab_0
pylint=2.7.2=py38h578d9bd_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h578d9bd_3
python=3.8.10=h49503c6_1_cpython
python-dateutil=2.8.1=py_0
python-louvain=0.15=pypi_0
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda10.2.89_cudnn7.6.5_0
pytz=2021.1=pyhd8ed1ab_0
qt=5.9.7=h5867ecd_1
rdflib=5.0.0=pypi_0
readline=8.1=h46c0cb4_0
regex=2021.4.4=py38h497a2fe_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
s3transfer=0.4.2=pyhd8ed1ab_0
scikit-learn=0.24.2=py38hdc147b9_0
scipy=1.6.3=py38h7b17777_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
setuptools=49.6.0=py38h578d9bd_3
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyh6c4a22f_0
smart_open=5.0.0=pyhd8ed1ab_0
sqlite=3.35.5=h74cdb3f_0
statsmodels=0.12.2=py38h5c078b8_0
tabulate=0.8.9=pyhd8ed1ab_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
toml=0.10.2=pyhd8ed1ab_0
torch-cluster=1.5.9=pypi_0
torch-geometric=1.7.0=pypi_0
torch-scatter=2.0.6=pypi_0
torch-sparse=0.6.9=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchvision=0.2.2=py_3
tornado=6.1=py38h497a2fe_1
tqdm=4.60.0=pyhd8ed1ab_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
unidecode=1.2.0=pyhd8ed1ab_0
urllib3=1.26.4=pyhd8ed1ab_0
werkzeug=2.0.0=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.12.1=py38h497a2fe_3
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
yarl=1.6.3=py38h497a2fe_1
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.4.9=ha95c52a_0
